+++ src/commands/ai/make_emoji.ts
@@ -1,10 +1,9 @@
 import type { Command } from '@commands';
 import { DURATION_TMP_EMOJI } from '@constants';
+import { DEFAULT_VALUE, EventReader } from '@utils/hugging_face';
 import { sleep } from 'bun';
 import { SlashCommandBuilder } from 'discord.js';
 import Jimp from 'jimp';
-
-import { DEFAULT_VALUE, EventReader } from './hugging_face.ts';
 
 function shrink_image(image: Jimp): Jimp {
     if (image.getWidth() > 256 && image.getHeight() > 256) {
